+++ CHANGES.rst
@@ -16,6 +16,7 @@ development
     - Refactor and improve library API and inline documentation
     - Provide new access methods ``fetch_st36xml`` and ``fetch_pdf``
     - Switch scraper from "mechanize" to "MechanicalSoup"
+    - Provide command line interface program ``dpmaregister``
 
 - Upgrade to Moment.js 2.20.0 re. CVE-2016-4055